+++ express-dao/src/main/java/edu/zut/cs/express/customer/dao/PayDao.java
@@ -1,7 +1,7 @@
-package edu.zut.cs.express.manager.dao;
+package edu.zut.cs.express.customer.dao;
 
 import edu.zut.cs.express.base.dao.GenericDao;
-import edu.zut.cs.express.manager.domain.Pay;
+import edu.zut.cs.express.customer.domain.Pay;
 
 public interface PayDao extends GenericDao<Pay, Long> {
