+++ express-dao/src/main/java/edu/zut/cs/express/customer/dao/User_usedao.java
@@ -1,7 +1,7 @@
-package edu.zut.cs.express.manager.dao;
+package edu.zut.cs.express.customer.dao;
 
 import edu.zut.cs.express.base.dao.GenericDao;
-import edu.zut.cs.express.manager.domain.User_use;
+import edu.zut.cs.express.customer.domain.User_use;
 
 public interface User_usedao extends GenericDao<User_use, Long> {
 }
