+++ express-dao/src/main/java/edu/zut/cs/express/customer/dao/user_registerDao.java
@@ -1,7 +1,7 @@
-package edu.zut.cs.express.manager.dao;
+package edu.zut.cs.express.customer.dao;
 
 import edu.zut.cs.express.base.dao.GenericDao;
-import edu.zut.cs.express.manager.domain.user_register;
+import edu.zut.cs.express.customer.domain.user_register;
 
 public interface user_registerDao extends GenericDao<user_register, Long> {
