+++ express-dao/src/test/java/edu/zut/cs/express/customer/dao/PayDaoTest.java
@@ -1,4 +1,4 @@
-package edu.zut.cs.express.manager.dao;
+package edu.zut.cs.express.customer.dao;
 
 import java.util.List;
 
@@ -6,8 +6,8 @@ import org.junit.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 
 import edu.zut.cs.express.base.dao.GenericDaoTestCase;
-import edu.zut.cs.express.manager.dao.PayDao;
-import edu.zut.cs.express.manager.domain.Pay;
+import edu.zut.cs.express.customer.dao.PayDao;
+import edu.zut.cs.express.customer.domain.Pay;
 
 public class PayDaoTest extends GenericDaoTestCase<Long, Pay, PayDao> {
