+++ express-dao/src/test/java/edu/zut/cs/express/customer/dao/User_usedaoTest.java
@@ -1,4 +1,4 @@
-package edu.zut.cs.express.manager.dao;
+package edu.zut.cs.express.customer.dao;
 
 import java.util.List;
 
@@ -6,8 +6,8 @@ import org.junit.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 
 import edu.zut.cs.express.base.dao.GenericDaoTestCase;
-import edu.zut.cs.express.manager.dao.User_usedao;
-import edu.zut.cs.express.manager.domain.User_use;
+import edu.zut.cs.express.customer.dao.User_usedao;
+import edu.zut.cs.express.customer.domain.User_use;
 
 public class User_usedaoTest extends GenericDaoTestCase<Long, User_use, User_usedao> {
