+++ express-dao/src/test/java/edu/zut/cs/express/customer/dao/user_orderDaoTest.java
@@ -1,4 +1,4 @@
-package edu.zut.cs.express.manager.dao;
+package edu.zut.cs.express.customer.dao;
 
 import java.util.List;
 
@@ -6,8 +6,8 @@ import org.junit.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 
 import edu.zut.cs.express.base.dao.GenericDaoTestCase;
-import edu.zut.cs.express.manager.dao.user_orderDAO;
-import edu.zut.cs.express.manager.domain.user_order;
+import edu.zut.cs.express.customer.dao.user_orderDAO;
+import edu.zut.cs.express.customer.domain.user_order;
 
 public class user_orderDaoTest extends GenericDaoTestCase<Long, user_order, user_orderDAO> {
