+++ express-dao/src/test/java/edu/zut/cs/express/customer/dao/user_registerDaoTest.java
@@ -1,4 +1,4 @@
-package edu.zut.cs.express.manager.dao;
+package edu.zut.cs.express.customer.dao;
 
 import java.util.List;
 
@@ -6,8 +6,8 @@ import org.junit.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 
 import edu.zut.cs.express.base.dao.GenericDaoTestCase;
-import edu.zut.cs.express.manager.dao.user_registerDao;
-import edu.zut.cs.express.manager.domain.user_register;
+import edu.zut.cs.express.customer.dao.user_registerDao;
+import edu.zut.cs.express.customer.domain.user_register;
 
 public class user_registerDaoTest extends GenericDaoTestCase<Long, user_register, user_registerDao> {
