+++ express-service/src/test/java/edu/zut/cs/express/manager/service/AdminServiceConfig.java
@@ -1,4 +1,4 @@
-package edu.zut.cs.express.admin.service;
+package edu.zut.cs.express.manager.service;
 
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
