+++ express-service/src/test/java/edu/zut/cs/express/manager/service/GroupManagerTest.java
@@ -1,7 +1,8 @@
-package edu.zut.cs.express.admin.service;
+package edu.zut.cs.express.manager.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
 
+import edu.zut.cs.express.admin.service.GroupManager;
 import edu.zut.cs.express.base.service.GenericTreeManagerTestCase;
 import edu.zut.cs.express.manager.domain.Group;
