+++ express-service/src/test/java/edu/zut/cs/express/manager/service/UserManagerTest.java
@@ -1,4 +1,4 @@
-package edu.zut.cs.express.admin.service;
+package edu.zut.cs.express.manager.service;
 
 import static org.junit.Assert.assertEquals;
 
@@ -10,6 +10,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 
+import edu.zut.cs.express.admin.service.UserManager;
 import edu.zut.cs.express.manager.domain.user;
 
 @RunWith(SpringJUnit4ClassRunner.class)
